pyproject.toml
[project]
name = "lemana"
version = "0.1.0"
description = ""
authors = [
    {name = "muhstefan"}
]
readme = "README.md"
requires-python = ">=3.13,<3.14"
dependencies = [
    "aiohttp (>=3.12.14,<4.0.0)",
    "gspread (>=6.2.1,<7.0.0)",
    "pytz (>=2025.2,<2026.0)",
    "python-dateutil (>=2.9.0.post0,<3.0.0)",
    "pyinstaller (>=6.14.2,<7.0.0)",
    "openpyxl (>=3.1.5,<4.0.0)",
    "selenium (>=4.34.2,<5.0.0)",
    "webdriver-manager (>=4.0.2,<5.0.0)",
    "beautifulsoup4 (>=4.13.4,<5.0.0)",
    "pandas (>=2.3.1,<3.0.0)",
    "astor (>=0.8.1,<0.9.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
